requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiodns==3.0.0
    # via aiohttp
aiofile==3.5.0
    # via -r requirements.in
aiohttp[speedups]==3.7.4.post0
    # via -r requirements.in
amqp==5.0.6
    # via kombu
asgiref==3.3.4
    # via django
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via aiohttp
backcall==0.2.0
    # via ipython
billiard==3.6.4.0
    # via celery
brotlipy==0.7.0
    # via
    #   -r requirements.in
    #   aiohttp
caio==0.8.1
    # via aiofile
cchardet==2.1.7
    # via aiohttp
celery-redbeat==2.0.0
    # via -r requirements.in
celery[redis]==5.1.0
    # via
    #   -r requirements.in
    #   celery-redbeat
    #   django-celery-beat
    #   django-celery-results
certifi==2020.12.5
    # via sentry-sdk
cffi==1.14.5
    # via
    #   brotlipy
    #   pycares
chardet==4.0.0
    # via aiohttp
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.1.6
    # via celery
click==7.1.2
    # via
    #   -r requirements.in
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
decorator==5.0.9
    # via ipython
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
dj-database-url==0.5.0
    # via -r requirements.in
django-celery-beat==2.2.0
    # via -r requirements.in
django-celery-results==2.0.1
    # via -r requirements.in
django-debreach==2.0.1
    # via -r requirements.in
django-import-export==2.5.0
    # via -r requirements.in
django-js-reverse==0.9.1
    # via -r requirements.in
django-log-request-id==1.6.0
    # via -r requirements.in
django-model-utils==4.1.1
    # via -r requirements.in
django-timezone-field==4.1.2
    # via django-celery-beat
django-webpack-loader==1.0.0
    # via -r requirements.in
django==3.2.3
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-import-export
    #   django-js-reverse
    #   django-log-request-id
    #   django-model-utils
    #   django-timezone-field
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
gunicorn==20.1.0
    # via -r requirements.in
idna==3.1
    # via yarl
ipython-genutils==0.2.0
    # via traitlets
ipython==7.23.1
    # via -r requirements.in
jedi==0.18.0
    # via ipython
kombu==5.1.0
    # via celery
markuppy==1.14
    # via tablib
matplotlib-inline==0.1.2
    # via ipython
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
odfpy==1.4.1
    # via tablib
openpyxl==3.0.7
    # via tablib
parso==0.8.2
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.18
    # via
    #   click-repl
    #   ipython
psutil==5.8.0
    # via -r requirements.in
psycopg2-binary==2.8.6
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pycares==4.0.0
    # via aiodns
pycparser==2.20
    # via cffi
pygments==2.9.0
    # via ipython
python-crontab==2.5.1
    # via django-celery-beat
python-dateutil==2.8.1
    # via
    #   celery-redbeat
    #   python-crontab
python-decouple==3.4
    # via -r requirements.in
pytz==2021.1
    # via
    #   celery
    #   django
    #   django-timezone-field
pyyaml==5.4.1
    # via tablib
redis==3.5.3
    # via
    #   celery
    #   celery-redbeat
sentry-sdk==1.1.0
    # via -r requirements.in
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
    #   tenacity
sqlparse==0.4.1
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.0.0
    # via django-import-export
tenacity==7.0.0
    # via celery-redbeat
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==3.10.0.0
    # via aiohttp
urllib3==1.26.4
    # via sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
websockets==9.0.2
    # via -r requirements.in
whitenoise==5.2.0
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
